# docs/swagger.yaml
basePath: /
host: localhost:4123
definitions:
  model.Application:
    properties:
      c_name:
        type: string
      created_at:
        type: string
      deployments:
        items:
          $ref: '#/definitions/model.ApplicationDeployment'
        type: array
      description:
        type: string
      e_name:
        type: string
      health_check_type:
        type: string
      health_check_url:
        type: string
      id:
        type: integer
      listen_port:
        type: integer
      repo:
        allOf:
        - $ref: '#/definitions/model.Repo'
        description: 修正外键关系定义
      repo_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ApplicationDeployment:
    properties:
      application:
        allOf:
        - $ref: '#/definitions/model.Application'
        description: 关联关系
      application_id:
        type: integer
      build_source:
        type: string
      build_template:
        $ref: '#/definitions/model.BuildTemplate'
      build_template_id:
        type: integer
      created_at:
        type: string
      deployment_template:
        $ref: '#/definitions/model.DeploymentTemplate'
      deployment_template_id:
        type: integer
      description:
        type: string
      environment:
        $ref: '#/definitions/model.Environment'
      environment_id:
        type: integer
      id:
        type: integer
      platform_credential_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ApplicationDeploymentRequest:
    properties:
      application_id:
        type: integer
      build_source:
        type: string
      build_template_id:
        type: integer
      deployment_template_id:
        type: integer
      description:
        type: string
      environment_id:
        type: integer
      platform_credential_id:
        type: integer
    type: object
  model.ApplicationDeploymentResponse:
    properties:
      application_id:
        type: integer
      build_source:
        type: string
      build_template_id:
        type: integer
      created_at:
        type: string
      deployment_template_id:
        type: integer
      description:
        type: string
      environment_id:
        type: integer
      id:
        type: integer
      platform_credential_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.BuildTemplate:
    properties:
      created_at:
        type: string
      creator:
        type: string
      dockerfile:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      pipeline:
        type: string
      repos:
        description: 关联Repo表的多对多关系
        items:
          $ref: '#/definitions/model.Repo'
        type: array
      updated_at:
        type: string
      updater:
        type: string
    type: object
...
  model.DeployTask:
    properties:
      created_at:
        type: string
      deployment_name:
        type: string
      env_id:
        type: integer
      finished_at:
        type: string
      git_ref:
        type: string
      harbor_project:
        type: string
      id:
        type: integer
      image_name:
        type: string
      image_tag:
        type: string
      jenkins_job_name:
        type: string
      k8s_cluster_id:
        description: 新增字段
        type: integer
      k8s_namespace:
        type: string
      log_path:
        type: string
      project_id:
        type: integer
      started_at:
        type: string
      status:
        description: PENDING, BUILDING, PUSHING, DEPLOYING, SUCCESS, FAILED
        type: string
      updated_at:
        type: string
    type: object
  model.DeploymentTemplate:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      display_name:
        type: string
      histories:
        description: 正确的关联历史记录关系
        items:
          $ref: '#/definitions/model.DeploymentTemplateHistory'
        type: array
      id:
        type: integer
      name:
        type: string
      parameters:
        type: string
      repos:
        description: 关联仓库表的多对多关系
        items:
          $ref: '#/definitions/model.Repo'
        type: array
      status:
        type: string
      template_type:
        type: string
      updated_at:
        type: string
      updater:
        type: string
      variables:
        type: string
      version:
        type: string
    type: object
  model.DeploymentTemplateHistory:
    properties:
      change_reason:
        type: string
      content:
        type: string
      created_at:
        type: string
      deployment_template_id:
        type: integer
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      modifier:
        type: string
      name:
        type: string
      parameters:
        type: string
      template_type:
        type: string
      variables:
        type: string
      version:
        type: string
    type: object
...
  model.Environment:
    properties:
      config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Repo:
    properties:
      applications:
        items:
          $ref: '#/definitions/model.Application'
        type: array
      c_name:
        type: string
      created_at:
        type: string
      deployment_templates:
        description: 添加与部署模板的多对多关联关系
        items:
          $ref: '#/definitions/model.DeploymentTemplate'
        type: array
      e_name:
        type: string
      id:
        type: integer
      platform:
        type: string
      repo_build_path:
        type: string
      repo_department:
        type: string
      repo_deploy_type:
        type: string
      repo_desc:
        type: string
      repo_language:
        type: string
      repo_manager:
        type: string
      repo_number:
        type: string
      repo_ssh_url:
        type: string
      repo_url:
        type: string
      templates:
        description: 添加与构建模板的多对多关联关系
        items:
          $ref: '#/definitions/model.BuildTemplate'
        type: array
      updated_at:
        type: string
    type: object
...
  types.PageResponse:
    properties:
      records: {}
      total:
        type: integer
    type: object
...
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Minimal OpenAPI spec for Zebra-CICD endpoints
  title: Zebra-CICD API
  version: 0.1.0
paths:
  /api/application/template:
    get:
      description: 根据应用服务ID获取部署配置列表
      parameters:
      - description: 应用服务ID
        in: query
        name: application_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationDeployment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取应用部署配置列表
      tags:
      - application-deployments
    post:
      consumes:
      - application/json
      description: 为应用服务创建部署配置
      parameters:
      - description: 部署配置信息
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationDeploymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建应用部署配置
      tags:
      - application-deployments
  /api/application/template/{id}:
    delete:
      description: 根据ID删除部署配置
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除应用部署配置
      tags:
      - application-deployments
    get:
      description: 根据部署配置ID获取详细信息
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取应用部署配置
      tags:
      - application-deployments
    put:
      consumes:
      - application/json
      description: 根据ID更新部署配置信息
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署配置信息
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationDeploymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用部署配置
      tags:
      - application-deployments
...
  /api/deploys:
    post:
      consumes:
      - application/json
      description: 创建一个新的部署任务，触发Jenkins构建并部署到K8s
      parameters:
      - description: 部署任务信息
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.DeployTask'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建部署任务
      tags:
      - deploys
  /api/deploys/{id}:
    get:
      description: 根据任务ID获取部署任务详情
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeployTask'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取部署任务
      tags:
      - deploys
...
  /api/repos/{id}/templates:
    get:
      description: 根据仓库ID获取所有关联的构建模板
      parameters:
      - description: 仓库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BuildTemplate'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取仓库关联的模板
      tags:
      - repos
...
  /api/templates/deployment:
    get:
      description: 获取所有部署模板的列表，支持按名称、类型、状态等条件查询
      parameters:
      - description: 部署模板名称
        in: query
        name: name
        type: string
      - description: 模板类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 创建人
        in: query
        name: creator
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 页数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/types.PageResponse'
                  - properties:
                      records:
                        items:
                          $ref: '#/definitions/model.DeploymentTemplate'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取部署模板列表
      tags:
      - deployment-templates
    post:
      consumes:
      - application/json
      description: 创建一个新的部署模板
      parameters:
      - description: 部署模板信息
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.DeploymentTemplate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DeploymentTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建部署模板
      tags:
      - deployment-templates
  /api/templates/deployment/{id}:
    delete:
      description: 根据部署模板ID删除部署模板（软删除）
      parameters:
      - description: 部署模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除部署模板
      tags:
      - deployment-templates
    get:
      description: 根据部署模板ID获取部署模板详情
      parameters:
      - description: 部署模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeploymentTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取部署模板
      tags:
      - deployment-templates
    put:
      consumes:
      - application/json
      description: 根据部署模板ID更新部署模板信息
      parameters:
      - description: 部署模板ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署模板信息
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.DeploymentTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeploymentTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新部署模板
      tags:
      - deployment-templates
  /api/templates/deployment/{id}/history:
    get:
      description: 根据部署模板ID获取模板的修改历史记录
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DeploymentTemplateHistory'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取部署模板历史记录
      tags:
      - deployment-templates
  /api/templates/deployment/{id}/repos:
    get:
      description: 根据部署模板ID获取所有关联的仓库列表
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Repo'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取关联的仓库列表
      tags:
      - deployment-templates
  /api/templates/deployment/{id}/repos/{repoId}:
    delete:
      description: 取消仓库与部署模板的关联
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      - description: 仓库ID
        in: path
        name: repoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 取消仓库和部署模板关联
      tags:
      - deployment-template-repo
    post:
      description: 将仓库与部署模板进行关联
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      - description: 仓库ID
        in: path
        name: repoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 关联仓库和部署模板
      tags:
      - deployment-template-repo
...
swagger: "2.0"